+++ sentry_sdk/api.py
@@ -1,5 +1,6 @@
 import inspect
 
+from sentry_sdk import tracing_utils
 from sentry_sdk._types import TYPE_CHECKING
 from sentry_sdk.hub import Hub
 from sentry_sdk.scope import Scope
@@ -238,11 +239,7 @@ def get_current_span(hub=None):
     """
     Returns the currently active span if there is one running, otherwise `None`
     """
-    if hub is None:
-        hub = Hub.current
-
-    current_span = hub.scope.span
-    return current_span
+    return tracing_utils.get_current_span(hub)
 
 
 def get_traceparent():
